backend/echochat/requirements.txt
fastapi
uvicorn
firebase-admin
google-cloud-firestore
boto3